requirements.txt
# Python Dependencies for NPK Sensor Monitor

# Serial communication
pyserial>=3.5

# Modbus protocol
minimalmodbus>=2.0.1

# MQTT client
paho-mqtt>=1.6.1

# Configuration management
PyYAML>=6.0

# Optional: Streamlit for local dashboard
streamlit>=1.28.0

# Optional: Data visualization for dashboard
plotly>=5.17.0
pandas>=2.0.0
